Now deleting boards by onLongPress
