+++ src/components/Boards/DeleteBoard.js
@@ -8,15 +8,10 @@ import {
   centeredView,
   modalView,
   buttonBlock,
-} from '../styles/styles';
-import {deleteBoard} from '../store/board/actions';
+} from '../../styles/styles';
+import {deleteBoard} from '../../store/board/actions';
 
 class DeleteBoard extends React.Component {
-  state = {
-    setModalVisible: false,
-    modalVisible: false,
-  };
-
   render() {
     return (
       <View style={centeredView}>
@@ -24,28 +19,21 @@ class DeleteBoard extends React.Component {
           animationType="slide"
           transparent={true}
           visible={this.props.isOpen || false}
-          onBackdropPress={() =>
-            this.setState({modalVisible: !this.state.modalVisible})
-          }
+          onBackdropPress={this.props.close}
           backgroundColor="rgba(0,0,0,0.7)">
           <View style={centeredView}>
             <View style={modalView}>
               <Text>Хотите удалить доску?</Text>
-              <View style={buttonBlock}>
+              <View style={[buttonBlock, {marginTop: 15}]}>
                 <TouchableOpacity
                   style={modalBtns}
                   onPress={() => {
-                    this.deleteBoard(this.props.boardId);
-                    // this.setState({modalVisible: !this.state.modalVisible});
+                    this.props.deleteBoard(this.props.boardId);
                     this.props.close();
                   }}>
                   <Icon name={'check'} color="#fff" size={20} />
                 </TouchableOpacity>
-                <TouchableOpacity
-                  style={modalBtns}
-                  onPress={() => {
-                    this.setState({modalVisible: !this.state.modalVisible});
-                  }}>
+                <TouchableOpacity style={modalBtns} onPress={this.props.close}>
                   <Icon name={'times'} color="#fff" size={20} />
                 </TouchableOpacity>
               </View>
